Extract extension methods from LoggerFactory class
This makes LoggerFactory's intellisense less cluttered.
